+++ PAINEL_CODIGO/TouchGFX/gui/include/gui/modoprova_screen/ModoProvaPresenter.hpp
@@ -19,12 +19,12 @@ public:
     virtual ~ModoProvaPresenter() {};
 
     virtual void updateRPMValue(int val);
-    virtual void updateSpeedValue(int val);
+    virtual void updateSpeedValue(float val);
     virtual void updateSOCValue(int val);
     virtual void updateFreioValue(int val);
     virtual void updateAceleradorValue(int val);
     virtual void hwButtonClicked(uint8_t buttonId);
-    virtual void updateTensaoHVValue(int val);
+    virtual void updateTensaoHVValue(float val);
     virtual void updateDistanciaValue(int val);
     virtual void updateTempAcumuladorValue(int val);
