+++ PAINEL_CODIGO/TouchGFX/gui/include/gui/modoprova_screen/ModoProvaView.hpp
@@ -13,7 +13,7 @@ public:
     virtual void tearDownScreen();
     virtual void updateSOC(int val);
     virtual void updateRPMValue(int val);
-    virtual void updateSpeedValue(int val);
+    virtual void updateSpeedValue(float val);
     virtual void updateFreio(int val);
     virtual void updateAcelerador(int val);
     virtual void updateTensaoHV(float val);
